api/requirements.txt
SQLAlchemy==1.4.32
uvicorn
psycopg2-binary
fastapi
google-auth
requests
pyjwt